win-prompt/execution/requirements.txt
google-generativeai
fastapi
uvicorn
python-dotenv
requests
firebase-admin
pydantic
python-multipart
